Silence warnings introduced in recent code merge.
git-svn-id: https://sourceforge.llnl.gov/svn/repos/samrai_project/trunk/MASTER@861 f45145c3-882e-4954-aa44-00140844b9ef

+++ source/test/performance/TreeCommunication/tcb.C
@@ -977,7 +977,7 @@ void simulateDataProcessing( const CommonTestSwitches &cts )
 */
 void my_usleep( size_t num_usec )
 {
-   size_t num_calls = static_cast<size_t>(num_usec * my_usleep_calls_per_usec);
+   size_t num_calls = static_cast<size_t>(double(num_usec) * my_usleep_calls_per_usec);
    for ( size_t i=0; i<num_calls; ++i ) {
       dummy();
    }
